requirements.txt
# Core ML & Transformers
transformers>=4.36.0
torch>=2.0.0
sentencepiece>=0.1.99
sentence-transformers>=2.2.2
scikit-learn>=1.1.3
numpy>=1.21.0

# UI
gradio>=4.15.0
requests>=2.31.0  
gradio_client>=0.8.1

# OCR & Image Processing
pytesseract>=0.3.10
Pillow>=10.0.0
pymupdf>=1.23.7
pdf2image>=1.16.3
sacremoses>=0.0.53

# Translation & Tokenization
accelerate>=0.27.0
safetensors>=0.3.3

# Google Cloud Vision Support
google-cloud-vision>=3.4.5
google-generativeai>=0.3.2